requirements.txt
discord.py
python-dotenv
vrchatapi